Refactor everything at once
What could go wrong ?
