+++ _sass/main.scss
@@ -20,9 +20,6 @@ $danger: $red;
 .pokemon-input {
     text-align: center;
 }
-.pokemon-level {
-    width: 3em;
-}
 
 .pokemon-choice-tile {
     height: 24em;
